correção de tabelas

+++ src/main/java/com/controle/api/controller/AdicionalController.java
@@ -4,6 +4,11 @@ import java.util.ArrayList;
 import java.util.List;
 import javax.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageImpl;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.web.PageableDefault;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -24,6 +29,8 @@ import com.controle.api.repository.AdicionalRepository;
 import com.controle.api.service.AdicionalService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
+
+import org.springframework.data.domain.Sort;
 
 @RestController
 @CrossOrigin(origins = "*")
@@ -47,14 +54,36 @@ public class AdicionalController {
         return ResponseEntity.status(HttpStatus.CREATED).body(adicionalService.save(adicionalInputDto, null));
     }
 	
-    @GetMapping
+   /* @GetMapping
     @ApiOperation(value="Busca lista de adicionais cadastrados")
     public ResponseEntity<List<AdicionalDto>> getListAdicionais(@RequestParam(required = false) Boolean status){
         List<Adicional> adicionais= new ArrayList<>();
         adicionais = adicionalService.buscaListaAdicional(status);        
         return ResponseEntity.status(HttpStatus.OK).body(adicionalMapper.toAdicionalListDto(adicionais));
                 
+    }*/
+	
+   /* @GetMapping
+    public ResponseEntity<Page<AdicionalDto>> getListAdicionais(@PageableDefault(page = 0, size = 2, sort = "id", direction = Sort.Direction.ASC) 
+    								Pageable pageable){ teste
+        return ResponseEntity.status(HttpStatus.OK).body(adicionalService.findAll(pageable));
     }
+	
+    */
+    @GetMapping
+    @ApiOperation(value="Busca lista de adicionais cadastrados")
+    public ResponseEntity<Page<AdicionalDto>> getListAdicionais(@RequestParam(required = false) Boolean status,
+    		@PageableDefault(page = 0, size = 2, sort = "id", direction = Sort.Direction.DESC)Pageable pageable	){
+        
+        Page<AdicionalDto> listaAdicionaisDto = adicionalService.buscaListaAdicional(status, pageable);
+        
+        return ResponseEntity.status(HttpStatus.OK).body(listaAdicionaisDto);
+                
+    }
+    
+         
+    
+	
     
     @GetMapping("/{id}")
     @ApiOperation(value="Busca adicional cadastrado pelo ID")
@@ -70,11 +99,27 @@ public class AdicionalController {
         return ResponseEntity.status(HttpStatus.OK).body(adicionalService.save(adicionalDto, adicionalUpdate.getId()));
     }
     
+    //@DeleteMapping("/{id}")
+   // @ApiOperation(value="Deleta adicional pelo id")
+	//public void delete(Adicional adicional) {
+   // 	adicionalRepository.delete(adicional);		
+	//}
+   
     @DeleteMapping("/{id}")
-    @ApiOperation(value="Deleta adicional pelo id")
-	public void delete(Adicional adicional) {
-    	adicionalRepository.delete(adicional);		
-	}
+    public ResponseEntity<Adicional> deleta(@PathVariable(value = "id") Long id){
+    	try {
+    		Adicional adicional =adicionalService.findById(id); 
+    		if (adicional!= null) {
+    			adicionalRepository.delete(adicional);
+    			return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    		}
+    		return ResponseEntity.status(HttpStatus.NOT_FOUND).build();   		
+    	} catch(DataIntegrityViolationException e) {
+    		
+    		return null;
+    	}
+    	
+    }
